@@ -1,24 +1,37 @@
-import { ApolloProvider, ApolloClient, InMemoryCache, createHttpLink } from '@apollo/client';
-import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import {
+  ApolloProvider,
+  ApolloClient,
+  InMemoryCache,
+  createHttpLink,
+} from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
+import React from "react";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 
+import Header from "./components/Header";
+import Footer from "./components/Footer";
+import Home from "./pages/Home";
 
-import Header from './components/Header';
-import Footer from './components/Footer';
-import Home from './pages/Home';
-
-import NoMatch from './pages/NoMatch';
-import SinglePost from './pages/SinglePost';
-import Profile from './pages/Profile';
-
-
+import NoMatch from "./pages/NoMatch";
+import SinglePost from "./pages/SinglePost";
+import Profile from "./pages/Profile";
 
 const httpLink = createHttpLink({
-  uri: '/graphql',
+  uri: "/graphql",
+});
+
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem("id_token");
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : "",
+    },
+  };
 });
 
 const client = new ApolloClient({
-  link: httpLink,
+  link: authLink.concat(httpLink),
   cache: new InMemoryCache(),
 });
 
@@ -30,22 +43,13 @@ function App() {
           <Header />
           <div className="container">
             <Routes>
-              <Route
-                path='/'
-                element={<Home />}
-              />
-            <Route path="/profile">
-              <Route path=":username" element={<Profile />} />
-                  <Route path="" element={<Profile />} />
+              <Route path="/" element={<Home />} />
+              <Route path="/profile">
+                <Route path=":username" element={<Profile />} />
+                <Route path="" element={<Profile />} />
               </Route>
-              <Route
-                path="/post"
-                element={<SinglePost />}
-              />
-              <Route
-                path="*"
-                element={<NoMatch />}
-              />
+              <Route path="/post/:id" element={<SinglePost />} />
+              <Route path="*" element={<NoMatch />} />
             </Routes>
           </div>
           <Footer />
