@@ -1,5 +1,5 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
+import React from "react";
+import { Link } from "react-router-dom";
 
 const PostList = ({ posts, title }) => {
   if (!posts.length) {
@@ -10,7 +10,7 @@ const PostList = ({ posts, title }) => {
     <div>
       <h3>{title}</h3>
       {posts &&
-        posts.map(post => (
+        posts.map((post) => (
           <div key={post._id} className="card mb-3">
             <p className="card-header">
               <Link
@@ -19,15 +19,15 @@ const PostList = ({ posts, title }) => {
                 className="text-light"
               >
                 {post.username}
-              </Link>{' '}
+              </Link>{" "}
               post on {post.createdAt}
             </p>
             <div className="card-body">
               <Link to={`/post/${post._id}`}>
                 <p>{post.postText}</p>
                 <p className="mb-0">
-                  Comments: {post.commentCount} || Click to{' '}
-                  {post.commentCount ? 'see' : 'start'} the discussion!
+                  Comments: {post.commentCount} || Click to{" "}
+                  {post.commentCount ? "see" : "start"} the discussion!
                 </p>
               </Link>
             </div>
