@@ -1,20 +1,13 @@
 const { Schema, model } = require("mongoose");
 
-const gameSchema = new Schema(
-  {
-    gamename: {
-      type: String,
-      required: true,
-      trim: true,
-    },
-    users: [
-        {
-          type: Schema.Types.ObjectId,
-          ref: 'User'
-        }
-      ],
-  }
-);
+const gameSchema = new Schema({
+  gamename: {
+    type: String,
+    required: true,
+    trim: true,
+    unique: true,
+  },
+});
 
 const Game = model("Game", gameSchema);
 
