@@ -32,8 +32,8 @@ const userSchema = new Schema(
     games: [
       {
         type: Schema.Types.ObjectId,
-        ref: 'Game'
-      }
+        ref: "Game",
+      },
     ],
     krossies: [
       {
