@@ -15,6 +15,10 @@ const typeDefs = gql`
     createdAt: String
     username: String
   }
+  type Game {
+    _id: ID
+    gamename: String
+  }
   type User {
     _id: ID
     username: String
@@ -24,17 +28,14 @@ const typeDefs = gql`
     krossies: [User]
     games: [Game]
   }
-  type Game {
-    _id: ID
-    gamename: String
-    users: [User]
-  }
+
   type Query {
     me: User
     users: [User]
     user(username: String!): User
     posts(username: String): [Post]
     post(_id: ID!): Post
+    games: [Game]
     game(gamename: String!): [User]
   }
   type Mutation {
